Main Menu now accessible from Login Screen among minor changes
Currently for debugging purposes, you do not need to enter a user/password to reach the main menu screen. There is some level of security implemented for later for example if someone enters the wrong credentials 3 times the button turns grey and is disabled. A place for the logo to go on the main menu screen was also added as a minor fix.