Create Dialing Activity, Implemented Fake Phone Call Menu Settings, Add Spinner functionality, Add more resource files
Sorry for the relatively large commit (I'll try in the future to create smaller commits), but this commit almost completes the Fake Phone Call Feature. Every setting that you can choose in the menu now works as it should with the exception of the voices and that's only because the recordings do not exist yet (the code to get them working exists though). Ending the fake phone call will now have an alert message asking the user to save the recording or not (and there are TODO's where those functions will eventually go)